scp_project/scpproject/requirements.txt
asgiref
attrs
awsebcli
bcrypt
blessed
boto3
boto
botocore
cached-property
cement
certifi
cffi
charset-normalizer
colorama
cryptography
Django==2.1.15
django-environ
django-storages
docker
docker-compose
dockerpty
docopt
future
idna
importlib-metadata
jmespath
jsonschema
paramiko
pathspec
pycparser
PyNaCl
pyrsistent
python-dateutil
pytz
PyYAML
requests
s3transfer
semantic-version
six
sqlparse
termcolor
texttable
typing_extensions
urllib3
wcwidth
websocket-client
zipp